pyproject.toml
[project]
name = "chuk-mcp-solver"
version = "0.1.0"
description = "General-purpose constraint and optimization solver MCP server"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "CHUK AI", email = "info@chuk.ai" }]
keywords = [
    "mcp",
    "model-context-protocol",
    "constraint-solver",
    "optimization",
    "or-tools",
    "cp-sat",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "chuk-mcp-server>=0.8",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "ortools>=9.10.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "types-pyyaml>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/chuk-ai/chuk-mcp-solver"
Documentation = "https://github.com/chuk-ai/chuk-mcp-solver#readme"
Repository = "https://github.com/chuk-ai/chuk-mcp-solver"
Issues = "https://github.com/chuk-ai/chuk-mcp-solver/issues"

[project.scripts]
chuk-mcp-solver = "chuk_mcp_solver.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chuk_mcp_solver"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
markers = [
    "asyncio: mark test as async",
    "network: mark test as requiring network access",
    "slow: mark test as slow running",
]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["src/chuk_mcp_solver"]
omit = ["tests/*", "examples/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
fail_under = 90

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = ["tests/", "examples/", "src/chuk_mcp_solver/server.py"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [".git", "__pycache__", "build", "dist", ".venv"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501", # line too long, handled by formatter
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
